Got a few things done while away on vacation, but not as much as I would have liked. Worked a bit on the path tool interface and started with a tree generation algorithm, which is no more than a stub at this point.

+++ src/editor/height_map.cpp
@@ -17,7 +17,7 @@ namespace ts
     HeightMap generate_height_map(std::uint32_t map_size, std::uint32_t cell_size, 
                                   float max_z, float variance)
     {
-      const std::size_t row_width = map_size + 1;
+      const std::uint32_t row_width = map_size + 1;
 
       HeightMap map(Vector2u(row_width, row_width), cell_size);
